chore(contracts): bench cold start and invocation
